Fix memory management bugs.
