+++ src/engine.ts
@@ -28,7 +28,6 @@ import {NamedTensorMap, NamedVariableMap, TensorContainer, TypedArray} from './t
 import * as util from './util';
 
 interface ScopeState {
-  keep: Tensor[];
   track: Tensor[];
   name?: string;
 }
@@ -62,7 +61,9 @@ export type MemoryInfo = {
   unreliable?: boolean;
 };
 
-export interface TimingInfo extends BackendTimingInfo { wallMs: number; }
+export interface TimingInfo extends BackendTimingInfo {
+  wallMs: number;
+}
 
 export class Engine implements TensorManager {
   // Public since optimizers will use it.
@@ -81,11 +82,12 @@ export class Engine implements TensorManager {
   // Keep Tensors that parallel the tapes.
   private activeScope: ScopeState;
   private scopeStack: ScopeState[];
+  private keepTensors: Tensor[] = [];
   private profiler: Profiler;
 
   constructor(private backend: KernelBackend, public safeMode: boolean) {
     // Create a default outer scope.
-    this.activeScope = {keep: [], track: []};
+    this.activeScope = {track: []};
     this.scopeStack = [this.activeScope];
     this.profiler = new Profiler(backend);
   }
@@ -228,7 +230,7 @@ export class Engine implements TensorManager {
           'Safe mode is ON. Enclose all tensor operations inside tf.tidy(): ' +
           'tf.tidy(() => {...}) to avoid memory leaks.');
     }
-    this.activeScope.keep.push(result);
+    this.keepTensors.push(result);
     return result;
   }
 
@@ -244,7 +246,7 @@ export class Engine implements TensorManager {
       this.gradientScopeCount++;
     }
 
-    const scopeInfo: ScopeState = {keep: [], track: []};
+    const scopeInfo: ScopeState = {track: []};
     if (name) {
       scopeInfo.name = name;
     }
@@ -264,7 +266,7 @@ export class Engine implements TensorManager {
       }
     }
 
-    let tensorsToKeep = this.activeScope.keep;
+    let tensorsToKeep = this.keepTensors.slice();
     const tensorsToTrackInParent = util.extractTensorsFromContainer(result);
     tensorsToKeep = tensorsToKeep.concat(tensorsToTrackInParent);
 
@@ -282,14 +284,16 @@ export class Engine implements TensorManager {
       }
     }
 
-    this.scopeStack.pop();
+    const oldScope = this.scopeStack.pop();
     this.activeScope = this.scopeStack.length === 0 ?
-        {keep: [], track: []} :
+        {track: []} :
         this.scopeStack[this.scopeStack.length - 1];
 
     // Track the current result in the parent scope.
     tensorsToTrackInParent.forEach(tensor => {
-      if (!util.isTensorInList(tensor, this.activeScope.keep)) {
+      // Only track the tensor if was allocated in the inner scope.
+      if (!util.isTensorInList(tensor, this.keepTensors) &&
+          util.isTensorInList(tensor, oldScope.track)) {
         this.track(tensor);
       }
     });
